修正:myRoot関数が整数を返すので、返り値の型をdoubleに変更

+++ main.c
@@ -1,6 +1,6 @@
 #include <stdio.h>
 
-int myRoot(double x) // x の平方根を返す関数。x >= 0 ではないときエラー表示を行い exit(1) とする。 
+double myRoot(double x) // x の平方根を返す関数。x >= 0 ではないときエラー表示を行い exit(1) とする。 
 {
   int y;
   if( x < 0 ){　// 0も許容範囲にするため条件を変更
